// admin/src/types/statistics/statisticsWidget.tsx
export interface Statistics {
    total_portfolios: number;
    total_portfolio_categories: number;
    total_portfolio_tags: number;
    total_portfolio_options: number;
    total_admins: number;
    total_users: number;
    total_media: number;
    total_posts: number;
    total_blog_categories: number;
    total_blog_tags: number;
    total_emails: number;
    new_emails: number;
    unanswered_emails: number;
    total_tickets: number;
    open_tickets: number;
    active_tickets: number;
    unanswered_tickets: number;
} 